test(mesh): the census grader goes; its value and action rules stay
tests/mesh/test_performed_checks_census_matches_the_code.py graded the
"Performed checks:" docstrings of validate_command and validate_input_frame
against their own AST -- a mirror of the code enforced against the code. Its
15 behavioural tests (bool refused, numpy bool, non-finite, envelope re-read,
teleop charset, override_code bounds) move verbatim to
tests/mesh/test_validate_command_value_and_action_rules.py; the two census
scans that also carried a value premise (bool refused, narrowed envelope
refuses) are covered by tests in the moved set.
